+++ lib/main/atom/onSaveHandler.ts
@@ -14,10 +14,14 @@ import errorView = require('./errorView'); ///ts:import:generated
 import debugAtomTs = require('./debugAtomTs'); ///ts:import:generated
 
 export function handle(event: { filePath: string; editor: AtomCore.IEditor }) {
+    // As a fall back to make sure we sync up in case of anything bad happening elsewhere.
     var textUpdated = parent.updateText({ filePath: event.filePath, text: event.editor.getText() });
 
-    if (atomConfig.compileOnSave) {
+    // Compile on save
+    parent.getProjectFileDetails({ filePath: event.filePath }).then(fileDetails => {
+        if (!fileDetails.project.compileOnSave) return;
+
         textUpdated.then(() => parent.emitFile({ filePath: event.filePath }))
             .then((res) => errorView.showEmittedMessage(res));
-    }
+    });
 }