+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ We only plan strictly document the breaking changes. The rest is optional.
 
 # v1
 * `format` -> `formatCodeOptions` https://github.com/TypeStrong/atom-typescript/issues/178
+* `compileOnSave` is now a tsconfig option : https://github.com/TypeStrong/atom-typescript/issues/187
 
 # Planned
-* No breaking changes yet. 
+* No breaking changes yet.